+++ src/game/states/beat.ts
@@ -1,6 +1,6 @@
 import { gameState, TransitionInformation, transitionsFrom } from "./common";
 
-transitionsFrom["team-selection"] = (): TransitionInformation => {
+transitionsFrom["init"] = (): TransitionInformation => {
   gameState("beat");
   return { remaining: 3, title: "Sheep in..." };
 };