+++ src/misc/buildInfo.ts
@@ -1,3 +1,3 @@
 export const REPO = "voces/fixus";
-export const VERSION = "11";
-export const BUILD = "2026-05-27T02:13:59.532Z";
+export const VERSION = "12";
+export const BUILD = "2026-05-28T06:55:13.222Z";